+++ src/plugins/plot/inspector/StackedPlotsInspectorViewProvider.js
